Use folder references for all groups and sort files/folders
