+++ passKit/Extensions/String+Utilities.swift
@@ -1,20 +1,17 @@
 //
-//  StringExtension.swift
+//  String+Utilities.swift
 //  passKit
 //
 //  Created by Yishi Lin on 2018/9/23.
 //  Copyright © 2018 Bob Sun. All rights reserved.
 //
 
-import Foundation
-
-public extension String {
-
-    var trimmed: String {
+extension String {
+    public var trimmed: String {
         return trimmingCharacters(in: .whitespacesAndNewlines)
     }
 
-    func stringByAddingPercentEncodingForRFC3986() -> String? {
+    public func stringByAddingPercentEncodingForRFC3986() -> String? {
         let unreserved = "-._~/?"
         var allowed = CharacterSet.alphanumerics
         allowed.insert(charactersIn: unreserved)
